more pain but almost done
